Rename "datasources" directory to "handlers"
